delete user groups before deleting user from system
